+++ src/Math/Vector4.h
@@ -37,8 +37,8 @@ namespace Raycer
 		Vector4& operator/=(const Vector4& v);
 		Vector4& operator/=(double s);
 
-		double get(uint index) const;
-		void set(uint index, double value);
+		double get(uint64_t index) const;
+		void set(uint64_t index, double value);
 		double length() const;
 		double lengthSquared() const;
 		void normalizeLength();
