+++ src/Raytracing/Textures/PerlinNoiseTexture.h
@@ -24,11 +24,11 @@ namespace Raycer
 		double getValue(const Vector2& texcoord, const Vector3& position) const override;
 		Vector3 getNormalData(const Vector2& texcoord, const Vector3& position, TextureNormalDataType& type) const override;
 
-		uint seed = 1;
+		uint32_t seed = 1;
 		Vector3 scale = Vector3(10.0, 10.0, 10.0);
 		Color baseColor = Color(1.0, 1.0, 1.0);
 		bool isFbm = false;
-		uint octaves = 4;
+		uint64_t octaves = 4;
 		double lacunarity = 2.0;
 		double persistence = 0.5;
